Crate wiwi_wasm

wiwi-wasm

Nicer, higher level APIs for working with JavaScript from Rust (WASM), mostly revolving around builder-style APIs

MSRV

Rust 1.85, which is, at the time of writing, only available on the beta/nightly release channels.

These docs have been built from commit 1609226b.

Modules

extern_crates

reflect

util

Common builder API utilities and stuffs

Structs

ExternAny

ExternBigint

bigint primitive type

ExternBigintNs

Global BigInt namespace

ExternNumber

number primitive type

ExternNumberNs

Global Number namespace

ExternNumberObject

Number wrapper object

ExternObject

object type

ExternReflectNs

Global Reflect namespace

ExternString

string primitive type

ExternStringNs

Global String namespace

ExternStringObject

String wrapper object

Functions

bigint

Get the global BigInt namespace object

number

Get the global Number namespace object

reflect

Get the global Reflect namespace object

Attribute Macros

wasm_bindgen

A list of all the attributes can be found here: https://rustwasm.github.io/docs/wasm-bindgen/reference/attributes/index.html