wiwi_wasm

Module util

Common builder API utilities and stuffs

Re-exports

pub use self::marker::*;

Modules

marker

Structs

Init

Marker struct for a field in the initialised state, optionally containing more state in the form of another type S, and a more “general” type that can be “matched upon” in implementations, in T

PhantomData

Zero-sized type used to mark things that “act like” they own a T.

Uninit

Marker struct for a field in the uninitialised state

Traits

InitStatus

Trait for marker structs to hold state about if a field in a builder is initialised or not

IsInit

Marker trait for marker structs that represent initialised state

IsUninit

Marker trait for marker structs that represent uninitialised state

Slot

Types that can be used safely for slots of type T