wiwi_wasm::util
Trait SlotWrite

pub unsafe trait SlotWrite<S>
where
    Self: Sized,
{
    type State: SlotRead<S>;

    // Required method
    unsafe fn write(self, slot: &mut S);
}

Types that can be safely written to slots of type T, just like [SlotWriteUnchecked], but can only be implemented on types for which it is correct to use

For example, if we have an API that is expecting a string, we would only implement this trait for string types, but we can still implement [SlotWriteUnchecked] for types that isn’t guaranteed to be a string, like ExternAny.

To implement this trait, implement [SlotWriteUnchecked] first, as that trait contains the actual implementation functionality. This trait is “only” a marker trait for those who implement [SlotWriteUnchecked], and can guarantee that all values of the type are valid for this slot.

Required Associated Types

type State: SlotRead<S>

Type to store in the state parameter (should almost always be just Self)

Required Methods

unsafe fn write(self, slot: &mut S)

Writes self into slot, doing as little work as needed

Work should be deferred to read, if possible.

Safety

Implementors must store a value into slot, that read can read back later. read is allowed to rely on the fact that something has been written.

It may be undesireable to call write twice on the same slot, but even with that, there is no strict requirement that callers only call write once.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.